only build lessons in build.js
